requirements.txt
charm-crypto>=0.43
